added conditionals for review routes

+++ app/api/review_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request
-from app.models import db, Review
+from app.models import db, Review, Product
 from flask_login import login_required, current_user
 from app.forms import ReviewForm
 
@@ -12,6 +12,16 @@ review_routes = Blueprint('reviews', __name__)
 def create_review(product_id):
     
     print('\n\n we are in the POST route')
+    #check if product exists
+    product = Product.query.get(product_id)
+    if not product:
+        return ({'message': "Product could't be found"})
+    
+    #check if user already has reviewed product
+    existing_review = Review.query.filter_by(product_id=product_id, user_id=current_user.id).first()
+    if existing_review:
+        return {"message": "User already has a review for this product"}, 403
+    
     form = ReviewForm()
     form['csrf_token'].data = request.cookies['csrf_token']
     if form.validate_on_submit():
@@ -19,7 +29,7 @@ def create_review(product_id):
             product_id=product_id,
             user_id=current_user.id,
             review=form.data['review'],
-            stars=form.data['stars']
+            stars=form.data['stars'],
         )
         db.session.add(new_review)
         db.session.commit()
@@ -31,6 +41,10 @@ def create_review(product_id):
 def get_product_reviews(product_id):
     print('\n\n we are in the get reviews route\n\n')
     reviews = Review.query.filter_by(product_id=product_id).all()
+
+    if not reviews:
+        return {'message': "Product couldn't be found"}, 404
+    
     return {'reviews': [review.to_dict() for review in reviews]}
 
 #Update a review
@@ -38,16 +52,23 @@ def get_product_reviews(product_id):
 @login_required
 def update_review(review_id):
     review = Review.query.get(review_id)
+
+    if not review:
+        return {'message': "Review couldn't be found"}, 404
+    
     if review.user_id != current_user.id:
-        return {'errors': 'Unauthroized'}, 403
+        return {'errors': 'Unauthorized'}, 403
     
     form = ReviewForm()
     form['csrf_token'].data = request.cookies['csrf_token']
+
     if form.validate_on_submit():
         review.review = form.data['review']
         review.stars = form.data['stars']
+
         db.session.commit()
         return review.to_dict(), 200
+    
     return {'errors': form.errors()}, 400
 
 #Delete a review
@@ -56,6 +77,18 @@ def update_review(review_id):
 @login_required
 def delete_review(review_id):
     review = Review.query.get(review_id)
+
+    if not review:
+        return {'message': "Review couldn't be found"}
     db.session.delete(review)
     db.session.commit()
-    return {'message': 'Review was successfully deleted'}, 200
+    return {'message': 'Successfully deleted'}, 200
+
+#Get all reviews of current user
+
+@review_routes.route('reviews/current', methods=['GET'])
+@login_required
+def get_user_reviews():
+    user_id = current_user.id
+    reviews = Review.query.filter_by(user_id=user_id).all()
+    return {'reviews': [review.to_dict() for review in reviews]}, 200
